scripts/server/site-optimizer/requirements.txt
click>=8.1.7
pydantic>=2.5.0
pandas>=2.1.0
python-dotenv>=1.0.0
google-analytics-data>=0.18.0
paramiko>=3.4.0
rich>=13.7.0
tabulate>=0.9.0
PyYAML>=6.0.1
